TST: more robust baseline testing data
The previous version of the tests were based on a function that was
the same in x and y. This meant that the tests were not testing
possible asymmetries in the data. This new version of the tests
removes that restriction and illustrates how to use NUFFT with
arbitrary functions, while ensuring that the results replicate Numpy's
FFT implementation.